requirements.txt
Flask==3.1.3
mysql-connector-python==9.5.0
python-dotenv==1.0.1
